separate Automationicon
